test(ReservationAcceptanceTest): 다이나믹 테스트 추가
